+++ quasar/src/qs/graphics/Line.hpp
@@ -34,7 +34,7 @@ namespace qs
 		/// \param y2 Y2 position.
 		/// \param thickness Sets the line thickness, defaulting to 1.
 		///
-		Line(const float x1, const float y1, const float x2, const float y2, const int thickness = 1) noexcept;
+		Line(const float x1, const float y1, const float x2, const float y2, const float thickness = 1.0f) noexcept;
 
 		///
 		/// Destructor.
@@ -50,7 +50,7 @@ namespace qs
 		/// \param y2 Y2 position.
 		/// \param thickness Sets the line thickness, defaulting to 1.
 		///
-		void create(const float x1, const float y1, const float x2, const float y2, const int thickness = 1) noexcept;
+		void create(const float x1, const float y1, const float x2, const float y2, const float thickness = 1.0f) noexcept;
 
 		///
 		/// Bind as active VA.
@@ -66,7 +66,7 @@ namespace qs
 		///
 		/// Line thickness.
 		///
-		int m_thickness;
+		float m_thickness;
 
 		///
 		/// VertexArray handle.
